+++ NEWS.md
@@ -7,6 +7,8 @@ plumber 1.0.0.9999 Development version
 
 ### New features
 
+* Guess OpenApi response content type from serializer (@meztez #684)
+
 * Passing `edit = TRUE` to `plumb_api()` will open the API source file (#699)
 
 * OpenAPI Specification can be set using a file path. (@meztez #696)